Remove Log reader support
The /_log endpoint is very cumbersome (and probably worthless) to use
programatically.  It's also going away in future versions of CouchDB, so why
bother supporting it?

I'll make the server use some more sane logging mechanism (i.e a file, or
syslog) in the future.  For now, just log to stderr.