Make CreateDB() return just an error
Returning just the DB, with a deferred error, is very inconvenient when
CreateDB() may return a kivik.StatusConflict, because the database already
exists. While it was possible to check for this state explicitly:

    if err := client.CreateDB(...).Err(); kivik.StatusCode(err) == kivik.StatusConflict { ... }

it meant that using the returned db object for other operations was
impossible any time that a conflict may occur, because that conflict would
then be deffered to the next database call.

So this change makes it explicit, that CreateDB() *only* creates the DB.