Zero out the iterator's curVal on each iteration
Go's json package doesn't always set all fields when unmarshalling, e.g.
a missing field. So not doing this can leak values from a previous
iterations to the next.

This is a bit of an hack using reflect, it might not work if the
interface value is not a pointer type.

Fixes #420